+++ src/Main.java
@@ -43,9 +43,9 @@ public class Main {
 		}
 
 		// Convert first and last binaries to decimal
-		int firstDecimal = Integer.parseInt(concatenatedFirstBits, 2);
-		int lastDecimal = Integer.parseInt(concatenatedLastBits, 2);
-		Integer [] arrDecimals = new Integer[]{firstDecimal, lastDecimal};
+		long firstDecimal = Long.parseLong(concatenatedFirstBits, 2);
+		long lastDecimal = Long.parseLong(concatenatedLastBits, 2);
+		Long [] arrDecimals = new Long[]{firstDecimal, lastDecimal};
 		System.out.println(arrDecimals);
 		System.out.println(firstDecimal);
 		System.out.println(lastDecimal);
